Forms: componentize response file field
