+++ projects/packages/forms/src/dashboard/components/response-view/style.scss
@@ -102,111 +102,15 @@
 	white-space: pre-wrap;
 }
 
-.jp-forms__inbox-response-data-value .email-field {
-
-	@include mixins.flex-center;
-	gap: 4px;
-}
-
 .response-country-flag {
 	font-size: 16px;
 	display: inline;
 	padding-right: 4px;
 	vertical-align: middle;
-}
 
-.response-country-flag svg {
-	fill: currentColor;
-	vertical-align: text-top;
-}
-
-.jp-forms__inbox-response-data-value .file-field {
-	margin-top: 4px;
-
-	.file-field__item {
-		display: flex;
-		justify-content: space-between;
-		gap: 12px;
-		--wp-admin-theme-color: #2271b1;
-		--wp-components-color-accent: var(--wp-admin-theme-color, #2271b1);
-	}
-
-	.file-field__icon {
-		width: 32px;
-		min-width: 32px;
-		height: 32px;
-		margin-top: 8px;
-		background-size: 18px;
-		border-radius: 50%;
-		background-color: #f6f6f7;
-		background-image: url(../../../contact-form/images/file-icons/txt.svg);
-		background-repeat: no-repeat;
-		background-position: center;
-
-		@media (min-width: #{ (breakpoints.$break-medium + 1) }) {
-			width: 40px;
-			min-width: 40px;
-			height: 40px;
-			background-size: 24px;
-			margin-top: 0;
-		}
-	}
-
-	.file-field__icon.icon-cal {
-		background-image: url(../../../contact-form/images/file-icons/cal.svg);
-	}
-
-	.file-field__icon.icon-html {
-		background-image: url(../../../contact-form/images/file-icons/html.svg);
-	}
-
-	.file-field__icon.icon-mp3 {
-		background-image: url(../../../contact-form/images/file-icons/mp3.svg);
-	}
-
-	.file-field__icon.icon-mp4 {
-		background-image: url(../../../contact-form/images/file-icons/mp4.svg);
-	}
-
-	.file-field__icon.icon-pdf {
-		background-image: url(../../../contact-form/images/file-icons/pdf.svg);
-	}
-
-	.file-field__icon.icon-png {
-		background-image: url(../../../contact-form/images/file-icons/png.svg);
-	}
-
-	.file-field__icon.icon-ppt {
-		background-image: url(../../../contact-form/images/file-icons/ppt.svg);
-	}
-
-	.file-field__icon.icon-sql {
-		background-image: url(../../../contact-form/images/file-icons/sql.svg);
-	}
-
-	.file-field__icon.icon-txt {
-		background-image: url(../../../contact-form/images/file-icons/txt.svg);
-	}
-
-	.file-field__icon.icon-xls {
-		background-image: url(../../../contact-form/images/file-icons/xls.svg);
-	}
-
-	.file-field__icon.icon-zip {
-		background-image: url(../../../contact-form/images/file-icons/zip.svg);
-	}
-
-	.file-field__info {
-		display: flex;
-		gap: 12px;
-	}
-}
-
-.jp-forms__inbox-response-data-value .image-select-field {
-
-	.image-select-field-button {
-		gap: variables.$grid-unit-10;
-		height: auto;
+	svg {
+		fill: currentColor;
+		vertical-align: text-top;
 	}
 }
 
@@ -255,6 +159,28 @@
 	}
 }
 
+
+// Styling override when loaded in modal on mobile
+.components-modal__content {
+
+	.jp-forms__inbox-response-title,
+	.jp-forms__inbox-response-subtitle,
+	.jp-forms__inbox-response-meta,
+	.jp-forms__inbox-response-data {
+		padding-left: 0;
+		padding-right: 0;
+	}
+
+	.jp-forms__inbox-file-preview-modal {
+		max-width: 90vw;
+		width: auto;
+		z-index: var(--jp-forms-z-modal); // Higher than parent modal
+	}
+}
+
+/*
+ * File preview
+ */
 .jp-forms__inbox-file-preview-shell {
 	max-height: 70vh;
 	overflow: hidden;
@@ -277,24 +203,6 @@
 	object-fit: contain;
 }
 
-// Styling override when loaded in modal on mobile
-.components-modal__content {
-
-	.jp-forms__inbox-response-title,
-	.jp-forms__inbox-response-subtitle,
-	.jp-forms__inbox-response-meta,
-	.jp-forms__inbox-response-data {
-		padding-left: 0;
-		padding-right: 0;
-	}
-
-	.jp-forms__inbox-file-preview-modal {
-		max-width: 90vw;
-		width: auto;
-		z-index: var(--jp-forms-z-modal); // Higher than parent modal
-	}
-}
-
 // Ensure backdrop is below the file preview modal but above the response modal
 .components-modal__screen-overlay:nth-of-type(2) {
 	z-index: var(--jp-forms-z-modal-backdrop);
@@ -308,6 +216,24 @@
 	flex-wrap: wrap;
 }
 
-.jp-forms__inbox-spinner.jp-forms__inbox-spinner {
+.jp-forms__inbox-file-spinner.jp-forms__inbox-file-spinner {
 	margin-top: 0;
+}
+
+
+/*
+ * Specific fields
+ */
+.jp-forms__inbox-response-data-value .email-field {
+
+	@include mixins.flex-center;
+	gap: 4px;
+}
+
+.jp-forms__inbox-response-data-value .image-select-field {
+
+	.image-select-field-button {
+		gap: variables.$grid-unit-10;
+		height: auto;
+	}
 }
